Add QUIT handling behaviour
When QUIT command is received, the server deleted the mails mark as deleted for the current user. Left to do: close the thread communication. Add few functions in ParserJSON, to mark a specific mail with a specific tag for a specific user, also to unmark all the deleted mails of a user, and to delete the marked mails. Also change the behaviour in transaction. Now not deleting on the arraylist but on the json file